+++ contracts/vm_api/db.cpp
@@ -18,6 +18,7 @@ int32_t db_store_i64_ex(uint64_t code, uint64_t scope, uint64_t table, uint64_t 
    return get_vm_api()->db_store_i64_ex(code, scope, table, payer, id, (const char*)data, len);
 }
 
+
 void db_update_i64(int32_t iterator, uint64_t payer, const void* data, uint32_t len) {
    get_vm_api()->db_update_i64( iterator, payer, (const char*)data, len);
 }
@@ -26,7 +27,7 @@ void db_remove_i64(int32_t iterator) {
    get_vm_api()->db_remove_i64(iterator);
 }
 
-void db_update_i64_ex( uint64_t scope, uint64_t payer, uint64_t table, uint64_t id, const char* buffer, uint32_t buffer_size ) {
+void db_update_i64_ex( uint64_t scope, uint64_t payer, uint64_t table, uint64_t id, const char* buffer, size_t buffer_size ) {
    get_vm_api()->db_update_i64_ex(scope, payer, table, id, buffer, buffer_size);
 }
 
@@ -38,11 +39,11 @@ int32_t db_get_i64(int32_t iterator, void* data, uint32_t len) {
    return get_vm_api()->db_get_i64(iterator, data, len);
 }
 
-int32_t db_get_i64_ex( int itr, uint64_t* primary, char* buffer, uint32_t buffer_size ) {
+int32_t db_get_i64_ex( int itr, uint64_t* primary, char* buffer, size_t buffer_size ) {
    return get_vm_api()->db_get_i64_ex( itr, primary, buffer, buffer_size );
 }
 
-const char* db_get_i64_exex( int itr, uint32_t* buffer_size ) {
+const char* db_get_i64_exex( int itr, size_t* buffer_size ) {
    return get_vm_api()->db_get_i64_exex( itr,  buffer_size);
 }
 
@@ -269,11 +270,11 @@ int32_t db_idx_long_double_end(uint64_t code, uint64_t scope, uint64_t table) {
    return get_vm_api()->db_idx_long_double_end(code, scope, table);
 }
 
-int32_t db_store_i256( uint64_t scope, uint64_t table, uint64_t payer, void* id, int size, const char* buffer, uint32_t buffer_size ) {
+int32_t db_store_i256( uint64_t scope, uint64_t table, uint64_t payer, void* id, int size, const char* buffer, size_t buffer_size ) {
    return get_vm_api()->db_store_i256(scope, table, payer, id, size, buffer, buffer_size);
 }
 
-void db_update_i256( int iterator, uint64_t payer, const char* buffer, uint32_t buffer_size ) {
+void db_update_i256( int iterator, uint64_t payer, const char* buffer, size_t buffer_size ) {
    return get_vm_api()->db_update_i256(iterator, payer, buffer, buffer_size);
 }
 
@@ -281,27 +282,27 @@ void db_remove_i256( int iterator ) {
    return get_vm_api()->db_remove_i256(iterator);
 }
 
-int db_get_i256( int iterator, char* buffer, uint32_t buffer_size ) {
+int db_get_i256( int iterator, char* buffer, size_t buffer_size ) {
    return get_vm_api()->db_get_i256(iterator, buffer, buffer_size);
 }
 
-int db_find_i256( uint64_t code, uint64_t scope, uint64_t table, void* id, uint32_t size ) {
+int db_find_i256( uint64_t code, uint64_t scope, uint64_t table, void* id, size_t size ) {
    return get_vm_api()->db_find_i256(code, scope, table, id, size);
 }
 
-int db_previous_i256( int iterator, void* primary, uint32_t size ) {
+int db_previous_i256( int iterator, void* primary, size_t size ) {
    return get_vm_api()->db_previous_i256(iterator, primary, size);
 }
 
-int db_next_i256( int iterator, void* primary, uint32_t size ) {
+int db_next_i256( int iterator, void* primary, size_t size ) {
    return get_vm_api()->db_next_i256(iterator, primary, size);
 }
 
-int db_upperbound_i256( uint64_t code, uint64_t scope, uint64_t table, void* id, uint32_t size ) {
+int db_upperbound_i256( uint64_t code, uint64_t scope, uint64_t table, void* id, size_t size ) {
    return get_vm_api()->db_upperbound_i256(code, scope, table, id, size);
 }
 
-int db_lowerbound_i256( uint64_t code, uint64_t scope, uint64_t table, void* id, uint32_t size ) {
+int db_lowerbound_i256( uint64_t code, uint64_t scope, uint64_t table, void* id, size_t size ) {
    return get_vm_api()->db_lowerbound_i256(code, scope, table, id, size);
 }
