+++ contracts/vm_api/system.cpp
@@ -37,11 +37,11 @@ uint32_t  now() {
    return (uint32_t)( current_time() / 1000000 );
 }
 
-void set_last_error(const char* error, uint32_t size) {
+void set_last_error(const char* error, size_t size) {
    get_vm_api()->set_last_error(error, size);
 }
 
-uint32_t get_last_error(char* error, uint32_t size) {
+size_t get_last_error(char* error, size_t size) {
    return get_vm_api()->get_last_error(error, size);
 }
